@@ -1,32 +1,23 @@
 <template>
+  
+  
   <div id="app">
-    <div id="nav">
-      <router-link to="/">Home</router-link> |
-      <router-link to="/about">About</router-link>
-    </div>
-    <router-view/>
+    <NavBar />    
+     <router-view />
+     <Footer />
   </div>
+    
 </template>
+<script>
 
+import NavBar from './components/NavBar.vue'
+import Footer from './components/Footer.vue'
+export default {
+  components:{
+    NavBar,Footer
+  }
+}
+</script>
 <style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-}
 
-#nav {
-  padding: 30px;
-}
-
-#nav a {
-  font-weight: bold;
-  color: #2c3e50;
-}
-
-#nav a.router-link-exact-active {
-  color: #42b983;
-}
 </style>
